requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
clipboard==0.0.4
crypto==1.4.1
cryptography==2.3.1
dominate==2.3.4
elasticsearch==6.3.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Naked==0.1.31
ndg-httpsclient==0.5.1
pipenv==2018.7.1
pkg-resources==0.0.0
pusher==2.0.1
pyasn1==0.4.4
pycparser==2.18
PyJWT==1.6.4
pymongo==3.7.1
pyOpenSSL==18.0.0
pyperclip==1.6.4
python-dateutil==2.7.3
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
shellescape==3.4.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
